requirements.txt
opencv-python
numpy
requests
pytesseract
Pillow
rich
dotenv
instaloader
facebook-scraper
lxml_html_clean
discord_webhook